Rename folder accordingly
