+++ src/courses/models/course-status.enum.ts
