+++ src/courses/models/course.repository.ts
@@ -1,7 +1,7 @@
 import { EntityRepository, Repository } from "typeorm";
 import { CourseEntity } from "./course.entity";
-import { CourseStatus } from '../models/course-status.enum'; 
-import { CreateCourseDto } from '../models/dto/create-course.dto';
+import { CourseStatus } from './course-status.enum'; 
+import { CreateCourseDto } from './dto/create-course.dto';
 
 @EntityRepository(CourseEntity)
 export class CourseRepository extends Repository<CourseEntity> {
